test(Fabric): add e2e test for issue 577 (#2816)
## Description

Check which example screens from issues/PRs can be used in e2e testing
for tests `Test550, ..., Test577` and implement them if possible for
Fabric.

### Test550
Skipped because the issue is present on the current version (custom
native header back button image). When it is resolved, we might create
an e2e test for it.

### Test556
Skipped because we can't check flickering nor non-deterministic white
screen bug.

### Test564
Skipped because the issue is still present on the current version
(headers overlapping when one navigatior has `headerTransparent: true`).
When it is resolved, we might create an e2e test for it.

### Test577
Test created. It checks whether attempting to close a `gestureEnabled:
false` modal on iOS doesn't break the modal's height allowing the
content underneath to be visible. The test is present only on iOS
because of differences in `modal` between iOS and Android.

## Changes

- add `Test577`
- add comments for test screen in `apps/src/tests/index.ts` with the
reason for (not) implementing e2e test for it
- add `e2e-utils.ts` with `decribeIfiOS` function that will skip running
the tests if they are run on Android

## Test code and steps to reproduce

CI

## Checklist

- [x] Ensured that CI passes