fix(Android,Paper): fix flicker during formsSheet enter transition on slow Android devices - not preloaded formSheet (#2820)
## Description

On old architecture we have "enter transition postponing" mechanism,
however it has been applied only to formsheets with fit to contents. It
turned out that on **very low-end** Android devices the issue affects
also not `fitToContents` sheets.


## Changes

This PR extends the enter transition postponing mechanism to all
`sheetAllowedDetents` values.

> [!important]
> There still remains problem though, when the sheet is preloaded before
navigating to it. The content wrapper receives the layout from RN,
however the screen never receives the frame, because it is not initially
attached to the container (and native layout triggered from screen stack
does not reach it).
>
> Then, when when the transition starts, content wrapper is already laid
out, however the screen is not & **FOR SOME WEIRD** reasons it receives
many frame updates during the transition - animated value updates are
interleaved with `Screen.onLayout` calls - resulting in flicker effect.

## Test code and steps to reproduce

Was not able to reproduce the issue on our end :/

## Checklist

- [ ] Ensured that CI passes