fix(Windows): fix building and crashing issues (#2826)
## Description

Fixes problems with building and crashes in Windows apps. Some
components are still WIP and we'll finish work on them in the future.

Resolves
https://github.com/software-mansion/react-native-screens/issues/2541.

## Changes

- add Windows-specific implementations of `ScreenContentWrapper` and
`ScreenFooter`,
- start work on stack header and subviews, searchbar, stack animations -
these components are not implemented yet
- change minimum version of `WindowsTargetPlatform` to match `WinUI`
requirements,

## Test code and steps to reproduce

Create new `react-native-windows` app with components from
`react-native-screens`, build and run (it might be necessary to run `npx
react-native run-windows` twice in order to install all required
packages).

---------

Co-authored-by: adrianryt <adrianryt3@gmail.com>