test(e2e, Fabric): add e2e tests for issue/PR examples 619..640  (#2850)
## Description

Check which example screens from issues/PRs can be used in e2e testing
for tests `Test619, ..., Test640` and implement them if possible for
Fabric.

### Test619
Skipped because we can't check components "jumping".

### Test624
Skipped because the change in PR
https://github.com/software-mansion/react-native-screens/pull/624 is
connected to internal JS prop. The test screen seems to be repurposed.

### Test640
Test created. It checks if `RefreshControl` works correctly with `modal`
(refresh instead of the modal closing).

## Changes

- add `Test640`
- add comments for every test screen from this PR in
`apps/src/tests/index.ts` with the reason for (not) implementing e2e
test for it

## Test code and steps to reproduce

CI

## Checklist

- [x] Ensured that CI passes