test(Windows): select react-native version for test project based on react-native-windows latest version (#2857)
## Description

Select `react-native version` for test project based on
`react-native-windows` latest version.

Sometimes, when new `react-native` version is released but
`react-native-windows` is not yet available for this version, CI might
fail because of the version mismatch. Now we fetch latest version of
`react-native-windows` from NPM in format `x.yy.zz` and we use `x.yy` to
determine latest supported version of `react-native`. Then we use it to
initialize new `react-native` project.

## Changes

- modify Windows CI

## Test code and steps to reproduce

Windows CI

## Checklist

- [x] Ensured that Windows CI passes