feat(iOS): add `unstable_accessibilityContainerViewIsModal` prop for `FullWindowOverlay` (#2854)
## Description


The prop defaults to `true` due to backward compatibility reasons.
I believe this should be changed with major version change.

There is no need to update Android interfaces, because we render plain
`View` there in place of `FullWindowOverlay`.

The prop is `unstable` since I haven't decided yet on whether
I want to expose completely new prop or make
`RNSFullWindowOverlayContainer` use value of
`RNSFullWindowOverlay.accessibilityViewIsModal`.

This would be possible, because it is regular prop on `View`
component. See:
https://reactnative.dev/docs/view#accessibilityviewismodal-ios

Currently I went with new prop and not repurposing
`accessibilityViewIsModal` on `RNSFullWindowOverlay` since I'm not that
familiar with the accessibility API & want to avoid focus request
collisions etc. & in case the original prop becomes utilised in the
future.


## Changes

Added new prop.

## Test code and steps to reproduce

`Test1096` + VoiceOver, when
`unstable_accessibilityContainerViewIsModal: false` you should be able
to access content under the FWO with voice over.

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes