fix(iOS,Fabric): `TouchableOpacity` does not work on screens with `headerTranslucent: true` on notchless iOS devices with older OS versions (#2858)
## Description

Fixes #2855

The issue is obvious, I'm staggered that we have not caught it much
earlier.
When `headerconfig.translucent == true` we lay out the `RNSScreenView`
underneath the navigation bar & keep sending non-zero content offset to
screen shadow node.

## Changes

Now, when `headerconfig.translucent == true` we send `contentOffsetY ==
0`.

## Test code and steps to reproduce

Added `Test2855`. Also should be reproducible on `Test2466` when you set
`headerTranslucent: true` on the screen with pressables.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes

---------

Co-authored-by: kligarski <63918941+kligarski@users.noreply.github.com>