chore(iOS, Fabric): add tests for header interactions (#2848)
## Description

This PR adds tests for #2809, #1589 and
https://github.com/software-mansion/react-native-screens/pull/2800.

## Changes

The main assumption is to test all permutation of
headerBackButtonMenuEnabled, headerBackButtonDisplayMode and content of
button (default - taken from previous screen headerTitle, custom from
headerBackTitle and styled when there are styles applied). I hope this
will help us tracking regression as this part of the code is a bit
confusing and caused us problem in the past.
Additionally there are added "custom" scenarios that test default
default behavior with default text and custom text (if the back button
changes it's appearance depending on available space and
[headerBackButtonDisplayMode](https://developer.apple.com/documentation/uikit/uinavigationitem/backbuttondisplaymode-swift.property)
specification).
The some of the last cases are set to failing, and this is known
disparity reported in #2809, we plan to fix it soon after, but wonted to
make sure we don't cause a regression.

**Remark:**
I did it in a way that after each test we're returning to the screen
with all test cases for 2809, this is due to the fact that scrolling
through list of all Test* takes a lot and reloading RN after each test
would cause huge overhead. The downside of it is that if any of the test
fails if will cascade fail all other.

## Test code and steps to reproduce

You can run test with: 
```
detox build --configuration ios.sim.debug
detox test --configuration ios.sim.debug e2e/issuesTests/Test2809.e2e.ts -l debug
````

## Checklist

- [ ] Ensured that CI passes