refactor(iOS): limit scope of `shouldUseCustomBackBarButtonItem` flag (#2867)
## Description / Changes

This is continuation of
https://github.com/software-mansion/react-native-screens/pull/2860 that
proposes small refactor around back button logic. This area is very
confusing and in my opinion this is due to flag
`shouldUseCustomBackBarButtonItem` that breaks the flow of other if's
and causes other props to not work. Not we contained this flag in `if
(visible)` which currently is the only case when backButtonItem might be
used after the changes from #2800.

## Test code and steps to reproduce

see `FabricExample/e2e/issuesTests/Test2809.e2e.ts` and
`apps/src/tests/Test2809/index.tsx`

## Checklist
- [ ] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>