test(e2e, Fabric): add e2e tests for issue/PR examples 642..645 (#2875)
## Description

Check which example screens from issues/PRs can be used in e2e testing
for tests `Test642, Test645` and implement them if possible for Fabric.

### Test642
Skipped because we can't check status bar visibility/style.

### Test645
Test created, only on iOS (`headerLargeTitle` is supported only on iOS).
It checks if `headerLargeTitle` isn't collapsed when changing screens. I
changed `Test645` screen to use `ScrollView` because it was used in
original issue.

## Changes

- add `Test645`
- change regular `View` to `ScrollView` in test screen `Test645`
- add comments for every test screen from this PR in
`apps/src/tests/index.ts` with the reason for (not) implementing e2e
test for it

## Test code and steps to reproduce

CI

## Checklist

- [x] Ensured that CI passes