fix(Fabric,Android): header subviews do not support dynamic content changes (#2910)
## Description

Fixes
https://github.com/software-mansion/react-native-screens/issues/2714 on
Android
Fixes
https://github.com/software-mansion/react-native-screens/issues/2815 on
Android

See #2905 for detailed description.

## Changes

Removed call to `RNSScreenStaceaderSubviewShadowNode.setSize` in
corresponding component descriptor.

It seems that we do not need to enforce node size from HostTree. Setting
appropriate content offset is enough for pressables to function
correctly (assuming that native layout **does not change size of any
subview**). I currently can't come up with any scenario where this would
happen.

## Test code and steps to reproduce

I've tested:

* [x] `Test2714` introduced in PR with iOS fixes -
https://github.com/software-mansion/react-native-screens/pull/2905
* [x] Pressables in header -
https://github.com/software-mansion/react-native-screens/pull/2466,
* [x] Header title truncation -
https://github.com/software-mansion/react-native-screens/pull/2325 (only
few cases, as the list is long there) & noticed a regression (not
related to this PR, described in comment below the PR description),
* [x] Insets with orientation change (Android) -
https://github.com/software-mansion/react-native-screens/pull/2756
* [x] https://github.com/software-mansion/react-native-screens/pull/2807
(on `TestHeaderTitle` with call to `setOptions` in `useLayoutEffect`)
* [x] `Test2811` -
https://github.com/software-mansion/react-native-screens/pull/2811
* [x] https://github.com/software-mansion/react-native-screens/pull/2812
(with snippet provided in that PR description)



## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes