fix(Android): fix formSheet sliding from top when it contains input with autofocus (#2909)
## Description

On API 36, formSheet that contains input with autoFocus would sometimes
slide from the top.

Even though `react-native-screens` used
`com.google.android.material:material:1.6.1`, `InsetsAnimationCallback`
(which is not present in `1.6.1`) is called (it probably comes from
`com.google.android.material:material:1.12.0` dependency in
`react-native-edge-to-edge`). This callback interferes with our
formSheet entering/closing animations:

1. Values used to calculate start and end position in
`InsetsAnimationCallback`'s `onPrepare` and `onStart` are usually
incorrect. This has become a significant issue since API 36: for some
reason, animation created in `ScreenStackFragment`'s `onCreateAnimator`
is sometimes incorrecly accounted for in calculating
`startTranslationY`. The value becomes negative which means that the
formSheet slides from the top.
2. Both `InsetsAnimationCallback` and formSheet animations use
`setTranslationY` and they are overwriting each other's changes.
`InsetsAnimationCallback`'s `onProgress` is called just after
animation's `setTranslationY` so it takes precedence.

We decided to replace the window insets animation callback which is set
in `BottomSheetBehavior`'s `onLayoutChild` with our own implementation.
Currently it is empty in order to bring back previous behavior.

FormSheet that contains input with autoFocus still does not behave as
intented - only rarely it gets expanded with the keyboard. This seems to
be a timing issue and needs more investigation. It might be necessary to
implement logic in new window insets animation callback.

Fixes
https://github.com/software-mansion/react-native-screens/issues/2895.

## Changes

- change `com.google.android.material:material` version to `1.12.0`
- replace `BottomSheetBehavior`'s `InsetsAnimationCallback` with empty
implementation
- add test screen `Test2895`

## Test code and steps to reproduce

Open `Test2895` screen in the example app and open the formSheet.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>