feat(bottom-tabs,iOS): add support for freezing tab screens on JS-tab changes (#5)
## Description

Add support for freezing tab screens on JS tab changes.

We can't just straight up freeze the component, because it will cause
the contents to disappear
during the transition. There is always an alternative solution with
`snapshot` taking, decided to try this out though.

This causes the component to be stateful. I believe it to be better
than other workarounds for this issue, however.

I have verified that this causes only shallow re-renders of the
`BottomTabsScreen` - neither parent nor children components are
re-rendered.


## Changes

The `BottomTabsScreen` listens for native component lifecycle events &
freezes react children only
when it is not focused & native component isn't visible.

## Test code and steps to reproduce

`TestBottomTabs`.

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Ensured that CI passes