test(e2e, Fabric): add e2e tests for issue/PR examples 702..726 (#2928)
## Description

Check which example screens from issues/PRs can be used in e2e testing
for tests `Test702, ..., Test726` and implement them if possible for
Fabric.

### Test702
Skipped because we can't check animation in any meaningful way.

### Test706
Skipped because we can't check font weight.

### Test713
Skipped because the issue is still present (`transitionStart/End` events
are not fired when leaving the screen due to a call from the JS side;
the decision was that [this will not be
fixed](https://github.com/software-mansion/react-native-screens/issues/713#issuecomment-732209901)
if there is no important use case for it).

### Test726
Test created, only on iOS (original issue was related to iOS). It checks
if the app freezes after going back from screen with autofocus input. It
is important to note that the issue was reopened with similar, but not
the same, bug.

## Changes

- add `Test726` e2e test
- add testIDs to `Test726` test screen
- add comments for every test screen from this PR in
`apps/src/tests/index.ts` with the reason for (not) implementing e2e
test for it

## Test code and steps to reproduce

CI

## Checklist

- [ ] Ensured that CI passes