chore: add search bar to Example app and use it in e2e tests (#2919)
## Description

Added search bar to Example app. It is useful for e2e testing - before
we would often scroll very far to get to the test screen we wanted - it
took a lot of time. Now, we use the search bar to find the screen
quickly.

### Why is hiding the keyboard necessary in
`selectTestScreen(screenName)` on Android?
This is necessary to make the CI pass. Currently, we're using `default`
Android image and `Pixel 2` in the GitHub Runner. `AOSP` keyboard
displays a banner that asks for access to contacts. The banner happens
to be just in front of the button to the test (first search result) when
using `Pixel 2`'s screen resolution. Clicking it instead of the button
behind would open a popup asking for permission to contacts. CI then
would crash with `Test Failed: No activities in stage RESUMED. Did you
forget to launch the activity. (test.getActivity() or similar)?`.


https://github.com/user-attachments/assets/4848bc35-6f4b-4343-a3ac-8ddd242b452e

## Changes

- add search bar to example app
- add e2e utility function `selectTestScreen(screenName)`
- use `selectTestScreen(screenName)` in e2e tests for issue/PR examples

## Test code and steps to reproduce

CI

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes