Fix animation during modal replace
The fix boils down to removing check whether `changeRootIndex ==
controllers.count`. It was introduced here:
https://github.com/software-mansion/react-native-screens/pull/502/.
It can be represented differently: `changeRootIndex ==
controllers.lastIndex + 1` & if I get this correctly it checks whether
the modal change in question takes place at the top of view hierarchy,
with purpose of not animating non visible changes.

It does not take account the replace operation though. In case of
replace, the `changeRootIndex == controllers.lastIndex`.

Another aspect is that I believe this part of the check is not
necessary. If some modal view controller is being dismissed, **all**
view controllers above it will also endup dismissed.
Therefore, there is no scenario where this check would be useful.

Hence, this commit ends up removing the check.