refactor(iOS, Fabric): move updateFormSheetPresentationStyle to updateProps (#2952)
## Description

Previously, `updateFormSheetPresentationStyle` was called from
`finalizeUpdates` every time it was called, which would run many times
when the sheet was being dragged. This impacted performance and was the
cause of the bug which was fixed in
https://github.com/software-mansion/react-native-screens/pull/2935 by
introducing a flag for setting initial detent. Now, we are checking
`updateMask` in `finalizeUpdates` method. If the props have changed, we
call `updateFormSheetPresentationStyle`. We don't need the flag anymore
so I removed it as well.

Thanks to @lodev09 for [reporting the problem and suggesting possible
solution](https://github.com/software-mansion/react-native-screens/pull/2935#issuecomment-2905783835).

## Changes

- remove `_sheetHasInitialDetentSet`
- check if props have changed using `updateMask` before calling
`updateFormSheetPresentationStyle`

## Test code and steps to reproduce

Run formSheet-related test screens in the example app, such as
`Test2543`, `TestFormSheet` (you can change `sheetInitialDetent` in the
file and save to see the update), `Test2877`.

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Ensured that CI passes