feat(iOS,stack): add skeleton code for new stack implementation (#30)
This is quite a big commit. It adds all the necessary skeleton code for
new stack implementation. The approach taken here is very similar to the
new implementation of bottom tabs.

ScreenStackHost is the component responsible for creating & hosting the
navigation controller. It forwards information (props & subviews) from
Fabric renderer to RNSStackController. It is also "event beat source"
for update flushing for the controller.

StackScreen is a component responsible for creation of child view
controller, which it be later managed by. It also hosts the user-defined
contents view. Later it'll also have responsibilites of passing back and
forward react information (props, events).

Added `TestScreenStack`

- [ ] Ensured that CI passes