Schedule removal transition on UI thread on notification from NativeProxy
We should be fine doing this, because the `notifyScreenRemoved` code is
run on JS thread BEFORE transaction execution is scheduled on UI thread.

This code works only under assumption, that the transaction is scheduled
in a SERIAL, no prioritized manner.

TOOD: We need to confirm that this is a case on Android.