feat: add support for `screenId` prop allowing for native screen recognition (#2967)
## Description

Add support for `screenId` prop on native side. At the moment we use
this value only in JS code.
It was brought to my attention, however, that other libs that wanna
integrate might want that
to be exposed also on native side. 

## Changes

Added `screenId` prop & exposed it on both platforms.

## Test code and steps to reproduce

Any example & add a breakpoint at setter method.

## Checklist

- [ ] Ensured that CI passes