fix(iOS): handle presenting custom modal when attempting to exit screen with preventNativeDismiss  (#2976)
## Description

When trying to use a transparent full-screen `Modal` from `react-native`
after attempting to close a screen that uses `usePreventRemove` hook,
sometimes the `Modal` would not attatch to `rootViewController` but to
`RNSScreen` - this would cause a visual bug and prevent any interaction
with the modal.

This is connected to how we handle `preventNativeDismiss`, introduced in
https://github.com/software-mansion/react-native-screens/pull/1773. We
create an `interactionController` for the transition and after a small
delay, we asynchronously dispatch a cancellation of the transition.

When `react-native`'s Modal tries to present itself, the host component
calls `presentViewController` on its controller (`RNSScreen`). Normally,
`RNSScreen` forwards the request up to its ancestor because presenting a
full-screen modal should be handled by a `rootViewController`. This
behavior is described in [UIKit
docs](https://developer.apple.com/documentation/uikit/uiviewcontroller/present(_:animated:completion:)?language=objc).

When the transition is still ongoing, `RNSScreen` is detached,
`parentViewController` of `RNSScreen` is `nil`. It can't pass the modal
up so it tries to present it but this obviously causes visual bugs and a
warning:

```
Presenting view controller <RCTFabricModalHostViewController: 0x1193a2000> 
from detached view controller <RNSScreen: 0x11931fc00> is not supported, 
and may result in incorrect safe area insets and a corrupt root presentation. 
Make sure <RNSScreen: 0x11931fc00> is in the view controller hierarchy 
before presenting from it. Will become a hard exception in a future release.
```

The workaround is to delay presenting the modal after the transition
completes but `transitionCoordinator` in `RNSScreen` is `nil` - this
might be due to the screen being detached from the stack that handles
the transition.

Instead, we use `reactSuperView` and `reactViewController` to access
`RNSNavigationController` and check if `transitionCoordinator != nil`.

Fixes
https://github.com/software-mansion/react-native-screens/issues/2125.
Includes reproduction from
https://github.com/software-mansion/react-native-screens/pull/2357 by
@adrianryt, @alduzy.

## Changes

- override `presentViewController` in `RNSScreen` to delay presenting
modals when in transition,
- add Test2125 test screen.

## Screenshots / GIFs


https://github.com/user-attachments/assets/a2b7a3f5-6d65-490e-82cf-bc89c99d6716

## Test code and steps to reproduce

Run `Test2125` from example app.

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Ensured that CI passes